todo_service/pyproject.toml
[tool.poetry]
name = "todo-service"
version = "0.1.0"
description = ""
authors = ["Vadim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
pydantic = "^2.5.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
alembic = "^1.12.1"
python-decouple = "^3.8"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pre-commit = "^3.5.0"
kafka = "^1.3.5"
aiokafka = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
